# Conflicts:
#	server/src/main/java/com/kofoos/api/entity/Category.java
#	server/src/main/java/com/kofoos/api/entity/DislikedMaterial.java
#	server/src/main/java/com/kofoos/api/entity/DislikedMaterialDetail.java
#	server/src/main/java/com/kofoos/api/entity/EditorProductsList.java
#	server/src/main/java/com/kofoos/api/entity/EditorRecommendationArticle.java
#	server/src/main/java/com/kofoos/api/entity/History.java
#	server/src/main/java/com/kofoos/api/entity/Image.java
#	server/src/main/java/com/kofoos/api/entity/Product.java
#	server/src/main/java/com/kofoos/api/entity/ProductMaterial.java
#	server/src/main/java/com/kofoos/api/entity/User.java
#	server/src/main/java/com/kofoos/api/entity/UserDislikesMaterial.java
#	server/src/main/java/com/kofoos/api/entity/WishlistFolder.java
#	server/src/main/java/com/kofoos/api/entity/WishlistItem.java

+++ server/src/main/java/com/kofoos/api/entity/WishlistFolder.java
@@ -14,7 +14,7 @@ public class WishlistFolder {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "wishlist_folder_id")
-    private Long id;
+    private int id;
 
     @Column(name = "name")
     private String name;
@@ -25,6 +25,7 @@ public class WishlistFolder {
 
     @OneToMany(mappedBy = "wishlistFolder")
     private List<WishlistItem> wishlistitems = new ArrayList<>();
+
 
     @Builder
     private WishlistFolder(String name, User user) {
